refactor(processor): drop the unreachable normalize_inputs typo pattern
normalalize_inputs. is the misspelling the missing extraction came from, and
no checkpoint key can start with it. Keeping it beside the correctly spelled
entry leaves two patterns for one intent, one of which never fires.

Also move the dataset-prefix regression test to
tests/processor/test_migrate_policy_normalization.py so it sits with the unit it
exercises; test_migration_detection.py covers
DataProcessorPipeline._is_processor_config.
